Fix error

+++ code/backend/twitter/api/models.py
@@ -32,15 +32,15 @@ class UserStats(models.Model):
 
 
 class Tweet(djongo_models.Model):
-	id = djongo_models.IntegerField(primary_key=True, db_column="_id")
-	tweet_id = djongo_models.IntegerField(db_column="id")
-	user = djongo_models.IntegerField()
+	id = djongo_models.BigIntegerField(primary_key=True, db_column="_id")
+	tweet_id = djongo_models.BigIntegerField(db_column="id")
+	user = djongo_models.BigIntegerField()
 	is_quote_status = djongo_models.BooleanField()
 	created_at = djongo_models.DateTimeField()
 	# quoted_status_id = djongo_models.IntegerField(null=True, blank=True)
 	in_reply_to_screen_name = djongo_models.TextField(null=True, blank=True)
-	in_reply_to_user_id = djongo_models.IntegerField(null=True, blank=True)
-	in_reply_to_status_id = djongo_models.IntegerField(null=True, blank=True)
+	in_reply_to_user_id = djongo_models.BigIntegerField(null=True, blank=True)
+	in_reply_to_status_id = djongo_models.BigIntegerField(null=True, blank=True)
 
 	class Meta:
 		managed = True
